Block until PointInTimeBuilder is obtained
