+++ plugins/command-manager/src/main/java/com/wazuh/commandmanager/CommandManagerPlugin.java
@@ -21,6 +21,7 @@ import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.IndexScopedSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.settings.SettingsFilter;
+import org.opensearch.common.unit.TimeValue;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.XContentParser;
@@ -76,7 +77,7 @@ public class CommandManagerPlugin extends Plugin
     public static final Integer JOB_PERIOD_MINUTES = 1;
     public static final Integer COMMAND_BATCH_SIZE = 2;
     public static final Long SEARCH_QUERY_TIMEOUT = 10L;
-    public static final Long PIT_KEEPALIVE_SECONDS = 30L;
+    public static final TimeValue PIT_KEEPALIVE_SECONDS = TimeValue.timeValueSeconds(60L);
 
     private static final Logger log = LogManager.getLogger(CommandManagerPlugin.class);
